Updates, - Order service changes: Total amount calculation - Caching implemented for Menu service - Transaction Logging implemented using Aspect implementation - New transactionlog table

+++ src/main/java/com/cafe/entity/Orders.java
@@ -21,7 +21,7 @@ public class Orders {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private Long tableId;
-    private Double totalAmount;
+    private Long totalAmount;
     private String status;
     private LocalDateTime createDt;
 }
